without_spring_framework/src/test/java/hellspawn287/dependencyinjection/without_spring_framework/without_spring_framework/controllers/SetterInjectedControllerTest.java
package hellspawn287.dependencyinjection.without_spring_framework.without_spring_framework.controllers;

import hellspawn287.dependencyinjection.without_spring_framework.without_spring_framework.services.GreetingServiceImpl;
import org.junit.Before;
import org.junit.Test;

import static org.junit.Assert.assertEquals;

public class SetterInjectedControllerTest {

    private SetterInjectedController setterInjectedController;

    @Before
    public void setUp() {
        this.setterInjectedController = new SetterInjectedController();
        this.setterInjectedController.setGreetingService(new GreetingServiceImpl());
    }

    @Test
    public void testGreetings() {
        assertEquals(GreetingServiceImpl.HELLO_ROBERT, setterInjectedController.sayHello());
    }
}